Ajustes menu

+++ PokerGames/Views/ListaJogadoresView/ListaJogadoresTableViewController.m
@@ -279,9 +279,8 @@
     return headerView;
 }
 
--(float)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section {
-    
-    return  44.0;
+- (CGFloat)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section {
+    return  44;
 }
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
