+++ PokerGames/Views/RankingCampeonatoView/RankingCampeonatoTableViewController.m
@@ -151,9 +151,8 @@
     return headerView;
 }
 
--(float)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section {
-    
-    return  44.0;
+- (CGFloat)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section {
+    return  44;
 }
